refactor oauth repository
