+++ README.md
@@ -34,18 +34,18 @@ ON PROGRESS
     - client registration api /client-registration (generate client_id and client_secret) [v]
     - Oauth2 redirect ui url see below
         ```
-        https://authorization-server.com/authorize?
+        https://authorization-server.com/oauth/authorize?
         response_type=code
         &client_id=oiMjpzj7ESt-JdcBnlhutyPC <- get from client
         &redirect_uri=https://www.oauth.com/playground/authorization-code.html <- get from client
         &scope=photo+offline_access
         &state=xJkMG8oJESzNzfyt <- get from client
         ```
-        - login ui []
+        - login ui [v]
         - forgot password ui []
     - exchange authorization code api /token (sharing with login user) []
     ```
-    POST https://authorization-server.com/token 
+    POST https://authorization-server.com/oauth/token 
 
     request:
     grant_type=authorization_code
